Refactor: Restructure project to root level and apply your feedback fixes
This commit implements a major project restructuring and addresses several points of your feedback to improve usability and the viewing experience.

Key Changes:

1.  **Project Restructuring**:
    - Moved all contents from the `matrix_movie_project/` subdirectory to the repository root. This includes the main Python script, `run.sh`, `README.md`, `requirements.txt`, and the `scenes/`, `utils/`, and `ascii_art/` directories.
    - The old `matrix_movie_project/` directory has been removed.
    - Updated all Python import paths within the project (e.g., `from matrix_movie_project.utils` to `from utils`) to reflect the new flattened structure.
    - Adjusted `sys.path` manipulations in scripts to ensure correct module resolution from the new root, both for the main movie and for standalone testing of scenes/utils.
    - Updated `README.md` and `run.sh` to align with the new file locations and ensure instructions are accurate.

2.  **`run.sh` Script Enhancements**:
    - Changed shebang to `#!/usr/bin/env bash`.
    - Improved Python interpreter detection logic, trying `python3` first, then `python`.
    - Ensured the script robustly changes to its own directory (now the project root) before execution.
    - Enhanced messages to you during script startup and for error conditions.

3.  **Text Display Fix (`renderer.py`)**:
    - Corrected newline handling in the `typing_print` function to resolve an issue where text could appear bunched up on a single line.

4.  **Code Rain Duration Adjustment**:
    - Reduced the duration of Matrix code rain animations (initial and inter-scene) in the main movie script to 2 seconds, making them brief "screen wipe" effects as requested.

5.  **Dependency Check in Main Script**:
    - Added a runtime check at the beginning of `matrix_resurrections_terminal_movie.py` to verify if the `rich` library is installed, providing a helpful message if it's missing.
    - Refined error messaging for missing internal project modules.

All changes have been tested, and the terminal movie now runs correctly with the new structure and applied fixes.
