--- matrix_movie_project/README.md
@@ -1,74 +1,0 @@
-# The Matrix Resurrections: Terminal Movie Experience
-
-Welcome, Operator. You've accessed a unique system diagnostic: a terminal-based "movie" that reimagines *The Matrix Resurrections* from the perspective of the Matrix system itself. Witness the unfolding events as a series of logs, process traces, AI debug messages, and cryptic system commentary, all within your terminal.
-
-This project is a nerdy, immersive experience for tech enthusiasts and Matrix fans, designed to simulate a system administrator watching the digital world react to Neo and Trinity's return.
-
-## Features
-
-*   **Purely Terminal-Based**: No GUIs, just pure text-mode output in a Linux-style terminal.
-*   **Automated Playback**: A non-interactive, timed "movie" that runs for approximately 10-12 minutes.
-*   **Matrix Code Rain**: Iconic green code rain animations at the start and as transitions between scenes.
-*   **System Log Aesthetic**: Output styled to mimic `dmesg`, `journalctl`, `strace`, `tcpdump`, and other familiar Linux tools, but with a Matrix twist.
-*   **Humorous & Cryptic AI Commentary**: Get insights from the Analyst AI and other system processes through their debug logs and internal messages.
-*   **Follows Matrix 4 Plot**: Key plot points of *The Matrix Resurrections* are represented from the machine's viewpoint across 8 distinct scenes.
-
-## Prerequisites
-
-*   **Python 3**: Ensure you have Python 3 installed (preferably Python 3.7+).
-*   **Rich Library**: The project uses the `rich` Python library for terminal formatting and animations.
-*   A terminal that supports ANSI escape codes and UTF-8 (most modern terminals do).
-
-## Installation & Running the Movie
-
-1.  **Clone the Repository**:
-    ```bash
-    git clone <repository_url> # Replace <repository_url> with the actual URL
-    cd matrix_movie_project
-    ```
-
-2.  **Set up a Virtual Environment (Recommended)**:
-    ```bash
-    python3 -m venv venv
-    source venv/bin/activate  # On Windows: venv\Scripts\activate
-    ```
-
-3.  **Install Dependencies**:
-    ```bash
-    pip install -r requirements.txt
-    ```
-
-4.  **Run the Movie**:
-    *   **Using the shell script (recommended for Linux/macOS)**:
-        Make sure your terminal window is maximized for the best experience!
-        ```bash
-        sh run.sh
-        ```
-        If you get a permission error, you might need to make it executable first: `chmod +x run.sh` (though the subtask should have handled this).
-
-    *   **Directly with Python**:
-        Make sure your terminal window is maximized!
-        ```bash
-        python3 matrix_resurrections_terminal_movie.py
-        ```
-        (Or `python matrix_resurrections_terminal_movie.py` if `python3` is not your command for Python 3).
-
-5.  **Enjoy the Show!**
-    Press `Ctrl+C` at any time to exit the movie.
-
-## Project Structure
-
-*   `matrix_resurrections_terminal_movie.py`: The main driver script.
-*   `run.sh`: Convenience script to launch the movie.
-*   `requirements.txt`: Python dependencies.
-*   `scenes/`: Contains individual Python scripts for each scene.
-*   `utils/`: Contains helper utilities, like the `renderer.py` for terminal effects.
-*   `ascii_art/`: Was planned for ASCII art, but thematic art was removed to enhance the raw terminal feel. Currently holds a `.gitkeep` file.
-
-## A Note on Style
-
-This project intentionally avoids complex ASCII art for logos or faces within the movie itself (post initial feedback) to maintain the authentic feel of observing a live terminal feed. The focus is on the text, the timing, and the simulated system environment.
-
----
-
-"We can't see it, but we're all trapped inside a routine. A loop." - Bugs, *The Matrix Resurrections*
